+++ src/effects.rs
@@ -1,0 +1,29 @@
+use bevy::prelude::*;
+
+pub struct EffectsPlugin;
+impl Plugin for EffectsPlugin {
+    fn build(&self, app: &mut App) {
+        app.add_systems(Update, menu_blink_system);
+    }
+}
+
+#[derive(Component)]
+pub struct DrawBlinkTimer(pub Timer);
+
+fn menu_blink_system(
+    mut commands: Commands,
+    time: Res<Time>,
+    mut query: Query<(Entity, &mut DrawBlinkTimer, &ComputedVisibility)>,
+) {
+    for (entity, mut timer, visibility) in query.iter_mut() {
+        timer.0.tick(time.delta());
+        if timer.0.finished() {
+            let new_visibility = if visibility.is_visible() {
+                Visibility::Hidden
+            } else {
+                Visibility::Visible
+            };
+            commands.entity(entity).insert(new_visibility);
+        }
+    }
+}